Added tests for all tasks

+++ BonusTask1/BonusTask1/extra-task-1.cpp
@@ -46,7 +46,7 @@ double to_float_hours(int hours, int minutes, int seconds)
 {
     assert((minutes >= 0) && (minutes < 60));
     assert((seconds >= 0) && (seconds < 60));
-    return static_cast<double>(hours + minutes / 60 + seconds / 3600);
+    return double(hours) + double(minutes)/60 + double(seconds) / 3600;
     /*
         Return the total number of hours in the specified number
         of hours, minutes, and seconds.
